代码提交，无重大版本更迭

+++ SpringBoot前沿技术栈/21_ShardingSphere5/21.2_SpringBoot3.x-ShardingSphere5/src/main/java/com/example/entity/EceUser.java
@@ -25,7 +25,10 @@ public class EceUser implements Serializable {
     @Serial
     private static final long serialVersionUID = -935366043629512096L;
 
-    private Integer id;
+    /**
+     * 主键 Id
+     */
+    private long id;
     /**
      * eceId
      */
